OPD project/Django app/requirements.txt
django
djangorestframework
django-cors-headers
pymysql
mysqlclient
cryptography
